Preserve classifier information if it is declared in the build
In this case, we can keep this information. Then, if there are ONLY
declarations to that one classified Jar, we know that this is the one
that might be needed in other places as well or that could be removed.

Cases where there are dependencies with different classifiers to
the same component are still fuzzy as we don't know which Jar file
belonged to which classifier. We could analyze the Jar name for this.
But we still only have one component in the Graph with all Jars.
So this does not fit the current model very well. And would need
special handling in more places.